requirements.txt
certifi
cffi
charset-normalizer
cryptography
Deprecated
docker
idna
packaging
pycparser
PyGithub
PyJWT
PyNaCl
requests
urllib3
websocket-client
wrapt
